+++ src/gemseo_calibration/measure.py
@@ -16,13 +16,21 @@
 from __future__ import annotations
 
 from typing import ClassVar
+from typing import Dict
 
-from gemseo.core.dataset import Dataset
 from gemseo.core.mdofunctions.mdo_function import MDOFunction
 from numpy import infty
 from numpy import nanmax
 from numpy import nanmin
 from numpy import ndarray
+
+DataType = Dict[str, ndarray]
+"""The type of data.
+
+The data are set as ``{variable_name: variable_values}``
+where ``variable_values`` is a 2D NumPy array
+whose rows are the samples and columns are the components of the variable.
+"""
 
 
 class CalibrationMeasure(MDOFunction):
@@ -35,13 +43,15 @@ class CalibrationMeasure(MDOFunction):
     """Whether to maximize the calibration measure."""
 
     def __init__(
-        self, output_name: str, name: str | None = None, f_type: str | None = None
+        self,
+        output_name: str,
+        name: str = "",
+        f_type: MDOFunction.FunctionType = MDOFunction.FunctionType.NONE,
     ) -> None:
-        # noqa: D205,D212,D415
         """
         Args:
             output_name: The name of the output to be taken into account by the measure.
-        """
+        """  # noqa: D205,D212,D415
         self.output_name = output_name
         super().__init__(None, name or self._compute_name(), f_type=f_type)
         self._lower_bound = -infty
@@ -57,7 +67,7 @@ class CalibrationMeasure(MDOFunction):
         """Return the name of the measure."""
         return f"{self.__class__.__name__}({self.output_name})"
 
-    def set_reference_data(self, reference_dataset: Dataset) -> None:
+    def set_reference_data(self, reference_dataset: DataType) -> None:
         """Define the reference input-output data set.
 
         Args:
@@ -76,7 +86,7 @@ class CalibrationMeasure(MDOFunction):
         self._lower_bound = min(data.min(), self._lower_bound)
         self._upper_bound = max(data.max(), self._upper_bound)
 
-    def __call__(self, model_dataset: Dataset) -> float:
+    def __call__(self, model_dataset: DataType) -> float:
         """Measure the (in)consistency between the model dataset and the reference one.
 
         Args:
@@ -88,10 +98,7 @@ class CalibrationMeasure(MDOFunction):
         raise NotImplementedError
 
     @staticmethod
-    def _compare_data(
-        data: ndarray,
-        other_data: ndarray,
-    ) -> ndarray:
+    def _compare_data(data: ndarray, other_data: ndarray) -> ndarray:
         """Compare two data arrays.
 
         Args: