+++ src/gemseo_calibration/measures/integrated_measure.py
@@ -15,12 +15,12 @@
 """A module to compute the integrated measure between two data sets."""
 from __future__ import annotations
 
-from gemseo.core.dataset import Dataset
 from numpy import interp
 from numpy import mean
 from numpy import trapz as integrate
 
 from gemseo_calibration.measure import CalibrationMeasure
+from gemseo_calibration.measure import DataType
 
 
 class IntegratedMeasure(CalibrationMeasure):
@@ -33,14 +33,13 @@ class IntegratedMeasure(CalibrationMeasure):
         self,
         output_name: str,
         mesh_name: str,
-        name: str | None = None,
-        f_type: str | None = None,
+        name: str = "",
+        f_type: CalibrationMeasure.FunctionType = CalibrationMeasure.FunctionType.NONE,
     ) -> None:
-        # noqa: D205 D212 D415
         """
         Args:
             mesh_name: The name of the 1D mesh.
-        """
+        """  # noqa: D205 D212 D415
         self.mesh_name = mesh_name
         self.__reference_mesh = None
         super().__init__(output_name, name=name, f_type=f_type)
@@ -48,8 +47,7 @@ class IntegratedMeasure(CalibrationMeasure):
     def _compute_name(self) -> str:
         return f"{self.__class__.__name__}({self.output_name};{self.mesh_name})"
 
-    def __call__(self, model_dataset: Dataset) -> float:
-        # noqa: D102
+    def __call__(self, model_dataset: DataType) -> float:  # noqa: D102
         model_data = model_dataset[self.output_name]
         model_mesh = model_dataset[self.mesh_name]
         self._update_bounds(model_data)
@@ -67,11 +65,9 @@ class IntegratedMeasure(CalibrationMeasure):
         )
 
     @property
-    def full_output_name(self) -> str:
-        # noqa: D102
+    def full_output_name(self) -> str:  # noqa: D102
         return f"{self.output_name}[{self.mesh_name}]"
 
-    def set_reference_data(self, reference_dataset: Dataset) -> None:
-        # noqa: D102
+    def set_reference_data(self, reference_dataset: DataType) -> None:  # noqa: D102
         self.__reference_mesh = reference_dataset[self.mesh_name]
         super().set_reference_data(reference_dataset)