+++ src/gemseo_calibration/post_processor.py
@@ -16,8 +16,9 @@
 from __future__ import annotations
 
 from gemseo.algos.opt_problem import OptimizationProblem
-from gemseo.core.dataset import Dataset
 from gemseo.post.opt_post_processor import OptPostProcessor
+
+from gemseo_calibration.measure import DataType
 
 
 class CalibrationPostProcessor(OptPostProcessor):
@@ -26,18 +27,17 @@ class CalibrationPostProcessor(OptPostProcessor):
     def __init__(
         self,
         opt_problem: OptimizationProblem,
-        reference_data: Dataset,
-        prior_model_data: Dataset,
-        posterior_model_data: Dataset,
+        reference_data: DataType,
+        prior_model_data: DataType,
+        posterior_model_data: DataType,
     ) -> None:
-        # noqa: D104, D205, D212, D415
         """
         Args:
             opt_problem: The optimization problem to run.
             reference_data: The reference data.
             prior_model_data: The model data before the calibration.
             posterior_model_data: The model data after the calibration.
-        """
+        """  # noqa: D205, D212, D415
         super().__init__(opt_problem)
         self._reference_data = reference_data
         self._prior_model_data = prior_model_data