+++ tests/post/data_versus_model/test_data_versus_model.py
@@ -23,9 +23,10 @@ from __future__ import annotations
 import pytest
 from gemseo.algos.parameter_space import ParameterSpace
 from gemseo.disciplines.analytic import AnalyticDiscipline
-from gemseo.utils.testing import image_comparison
+from gemseo.utils.testing.helpers import image_comparison
 from gemseo_calibration.calibrator import CalibrationMeasure
 from gemseo_calibration.post.data_versus_model.post import DataVersusModel
+from gemseo_calibration.post.factory import CalibrationPostFactory
 from gemseo_calibration.scenario import CalibrationScenario
 from numpy import array
 
@@ -43,7 +44,7 @@ def calibration_scenario() -> CalibrationScenario:
     reference.set_cache_policy("MemoryFullCache")
     reference.execute({"x": array([1.0])})
     reference.execute({"x": array([2.0])})
-    reference_data = reference.cache.export_to_dataset(by_group=False)
+    reference_data = reference.cache.to_dataset().to_dict_of_arrays(False)
 
     calibration = CalibrationScenario(
         model,
@@ -80,9 +81,26 @@ def test_plot(
     pyplot_close_all,
 ):
     """Test images created by DataVersusModel._plot against references."""
-    DataVersusModel(
+    CalibrationPostFactory().create(
+        "DataVersusModel",
         calibration_scenario.formulation.opt_problem,
         calibration_scenario.calibrator.reference_data,
         calibration_scenario.prior_model_data,
         calibration_scenario.posterior_model_data,
-    )._plot(**kwargs)
+    ).execute(save=False, **kwargs)
+
+
+def test_factory_plot(calibration_scenario):
+    """Check CalibrationPostFactory.execute()."""
+    factory = CalibrationPostFactory()
+    post = factory.execute(
+        calibration_scenario.formulation.opt_problem,
+        calibration_scenario.calibrator.reference_data,
+        calibration_scenario.prior_model_data,
+        calibration_scenario.posterior_model_data,
+        "DataVersusModel",
+        output="y",
+        save=False,
+    )
+    assert isinstance(post, DataVersusModel)
+    assert factory.executed_post[-1] == post