+++ tests/post/test_factory.py
@@ -18,14 +18,13 @@ from __future__ import annotations
 from pathlib import Path
 
 import pytest
-from gemseo.problems.analytical.rosenbrock import Rosenbrock
 from gemseo_calibration.post.factory import CalibrationPostFactory
 
 DATA = Path(__file__).parent / ".." / "data"
 
 
 @pytest.fixture
-def post_factory(monkeypatch):  # type: (...) -> CalibrationPostFactory
+def post_factory(monkeypatch) -> CalibrationPostFactory:
     """The factory of post-processors dedicated to calibration."""
     monkeypatch.setenv("GEMSEO_PATH", DATA)
     return CalibrationPostFactory()
@@ -33,32 +32,10 @@ def post_factory(monkeypatch):  # type: (...) -> CalibrationPostFactory
 
 def test_init(post_factory):
     """Check that the factory is correctly initialized."""
-    assert (
-        post_factory.factory._Factory__base_class.__name__ == "CalibrationPostProcessor"
-    )
-
-
-def test_create(post_factory):
-    """Check that a post-processor is correctly created."""
-    post = post_factory.create(Rosenbrock(), 1, 2, 3, "NewCalibrationPostProcessor")
-    assert post.__class__.__name__ == "NewCalibrationPostProcessor"
-    assert post._reference_data == 1
-    assert post._prior_model_data == 2
-    assert post._posterior_model_data == 3
-
-
-def test_execute(post_factory):
-    """Check that a post-processor is correctly executed.
-
-    Args:
-        factory (CalibrationPostFactory): A factory
-            to post-process calibration scenarios.
-    """
-    post = post_factory.execute(Rosenbrock(), 1, 2, 3, "NewCalibrationPostProcessor")
-    assert post.executed
+    assert post_factory._CLASS.__name__ == "CalibrationPostProcessor"
 
 
 def test_posts(post_factory):
     """Check that a post-processor is correctly executed."""
-    assert "NewCalibrationPostProcessor" in post_factory.posts
-    assert "OptPostProcessor" not in post_factory.posts
+    assert "DataVersusModel" in post_factory.posts
+    assert "OptHistoryView" not in post_factory.posts