+++ CHANGELOG.rst
@@ -25,6 +25,17 @@ The format is based on
 and this project adheres to
 `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+Version 2.0.0 (June 2023)
+*************************
+
+Update to GEMSEO 5.
+
+Changed
+-------
+
+- Data are ``dict[str, ndarray]`` objects as ``{variable_name: variable_values}`` instead of :class:`~gemseo.datasets.dataset.Dataset`.
+- Use ``""`` as empty value of ``str`` and ``str | Path`` arguments, instead of ``"None"``.
+- :class:`.CalibrationMeasure`: the type of ``f_type`` is :attr:`~gemseo.core.mdofunctions.mdo_function.MDOFunction.FunctionType`.
 
 Version 1.0.0 (July 2022)
 *************************